+++ internal/config/state.go
@@ -15,10 +15,7 @@
 package config
 
 import (
-	"errors"
 	"fmt"
-	"log/slog"
-	"os"
 	"path/filepath"
 	"regexp"
 	"strings"
@@ -220,52 +217,4 @@ func isValidImage(image string) bool {
 	ref, tag := parseImage(image)
 
 	return ref != "" && tag != ""
-}
-
-// ReadResponse reads the library state from configure-response.json.
-//
-// The response file is removed afterwards.
-func ReadResponse(contentLoader func(data []byte, state *LibraryState) error, jsonFilePath string) (*LibraryState, error) {
-	data, err := os.ReadFile(jsonFilePath)
-	defer func() {
-		if err := os.Remove(jsonFilePath); err != nil {
-			slog.Warn("fail to remove file", slog.String("name", jsonFilePath), slog.Any("err", err))
-		}
-	}()
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response file, path: %s, error: %w", jsonFilePath, err)
-	}
-
-	libraryState := &LibraryState{}
-
-	if err := contentLoader(data, libraryState); err != nil {
-		return nil, fmt.Errorf("failed to load file, %s, to state: %w", jsonFilePath, err)
-	}
-
-	return libraryState, nil
-}
-
-// WriteLibrarianState writes the given librarian state to a yaml file.
-func WriteLibrarianState(contentParser func(state *LibrarianState) ([]byte, error), state *LibrarianState, yamlFilePath string) (err error) {
-	yamlFile, err := os.Create(yamlFilePath)
-	if err != nil {
-		return fmt.Errorf("failed to create librarian state file, path: %s, error: %w", yamlFilePath, err)
-	}
-
-	defer func() {
-		if err = errors.Join(err, yamlFile.Close()); err != nil {
-			err = fmt.Errorf("WriteLibrarianState(), file path: %s, error: %w", yamlFilePath, err)
-		}
-	}()
-
-	data, err := contentParser(state)
-	if err != nil {
-		return fmt.Errorf("failed to convert state to bytes: %w", err)
-	}
-
-	if _, err := yamlFile.Write(data); err != nil {
-		return fmt.Errorf("failed to write librarian state file, path: %s, error: %w", yamlFilePath, err)
-	}
-
-	return nil
 }
