+++ internal/librarian/state_test.go
@@ -16,9 +16,14 @@ package librarian
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
+	"fmt"
+	"gopkg.in/yaml.v3"
+	"io"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -333,4 +338,242 @@ func TestPopulateServiceConfig(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestReadResponseJSON(t *testing.T) {
+	t.Parallel()
+	contentLoader := func(data []byte, state *config.LibraryState) error {
+		return json.Unmarshal(data, state)
+	}
+	for _, test := range []struct {
+		name         string
+		jsonFilePath string
+		wantState    *config.LibraryState
+	}{
+		{
+			name:         "successful-unmarshal",
+			jsonFilePath: "../../testdata/successful-unmarshal-libraryState.json",
+			wantState: &config.LibraryState{
+				ID:                  "google-cloud-go",
+				Version:             "1.0.0",
+				LastGeneratedCommit: "abcd123",
+				APIs: []*config.API{
+					{
+						Path:          "google/cloud/compute/v1",
+						ServiceConfig: "example_service_config.yaml",
+					},
+				},
+				SourcePaths:   []string{"src/example/path"},
+				PreserveRegex: []string{"example-preserve-regex"},
+				RemoveRegex:   []string{"example-remove-regex"},
+			},
+		},
+		{
+			name:         "empty libraryState",
+			jsonFilePath: "../../testdata/empty-libraryState.json",
+			wantState:    &config.LibraryState{},
+		},
+		{
+			name:      "invalid_file_name",
+			wantState: nil,
+		},
+		{
+			name:         "invalid content loader",
+			jsonFilePath: "../../testdata/invalid-contentLoader.json",
+			wantState:    nil,
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			tempDir := t.TempDir()
+			if test.name == "invalid_file_name" {
+				filePath := filepath.Join(tempDir, "my\x00file.json")
+				_, err := readResponse(contentLoader, filePath)
+				if err == nil {
+					t.Error("readResponse() expected an error but got nil")
+				}
+
+				if g, w := err.Error(), "failed to read response file"; !strings.Contains(g, w) {
+					t.Errorf("got %q, wanted it to contain %q", g, w)
+				}
+
+				return
+			}
+
+			if test.name == "invalid content loader" {
+				invalidContentLoader := func(data []byte, state *config.LibraryState) error {
+					return errors.New("simulated Unmarshal error")
+				}
+				dst := fmt.Sprintf("%s/copy.json", os.TempDir())
+				if err := copyFile(dst, test.jsonFilePath); err != nil {
+					t.Error(err)
+				}
+				_, err := readResponse(invalidContentLoader, dst)
+				if err == nil {
+					t.Errorf("readResponse() expected an error but got nil")
+				}
+
+				if g, w := err.Error(), "failed to load file"; !strings.Contains(g, w) {
+					t.Errorf("got %q, wanted it to contain %q", g, w)
+				}
+				return
+			}
+
+			// The response file is removed by the readResponse() function,
+			// so we create a copy and read from it.
+			dstFilePath := fmt.Sprintf("%s/copy.json", os.TempDir())
+			if err := copyFile(dstFilePath, test.jsonFilePath); err != nil {
+				t.Error(err)
+			}
+
+			gotState, err := readResponse(contentLoader, dstFilePath)
+
+			if err != nil {
+				t.Fatalf("readResponse() unexpected error: %v", err)
+			}
+
+			if diff := cmp.Diff(test.wantState, gotState); diff != "" {
+				t.Errorf("Response library state mismatch (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestWriteLibrarianState(t *testing.T) {
+	t.Parallel()
+	for _, test := range []struct {
+		name  string
+		state *config.LibrarianState
+	}{
+		{
+			name: "successful-marshaling-librarianState-yaml",
+			state: &config.LibrarianState{
+				Image: "v1.0.0",
+				Libraries: []*config.LibraryState{
+					{
+						ID:                  "google-cloud-go",
+						Version:             "1.0.0",
+						LastGeneratedCommit: "abcd123",
+						APIs: []*config.API{
+							{
+								Path:          "google/cloud/compute/v1",
+								ServiceConfig: "example_service_config.yaml",
+							},
+						},
+						SourcePaths: []string{
+							"src/example/path",
+						},
+						PreserveRegex: []string{
+							"example-preserve-regex",
+						},
+						RemoveRegex: []string{
+							"example-remove-regex",
+						},
+					},
+					{
+						ID:      "google-cloud-storage",
+						Version: "1.2.3",
+						APIs: []*config.API{
+							{
+								Path:          "google/storage/v1",
+								ServiceConfig: "storage_service_config.yaml",
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name:  "empty-librarianState-yaml",
+			state: &config.LibrarianState{},
+		},
+		{
+			name:  "invalid_file_name",
+			state: &config.LibrarianState{},
+		},
+		{
+			name:  "invalid content parser",
+			state: &config.LibrarianState{},
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			tempDir := t.TempDir()
+			contentParser := func(state *config.LibrarianState) ([]byte, error) {
+				return yaml.Marshal(state)
+			}
+			if test.name == "invalid_file_name" {
+				filePath := filepath.Join(tempDir, "my\x00file.yaml")
+				err := writeLibrarianState(contentParser, test.state, filePath)
+				if err == nil {
+					t.Errorf("writeLibrarianState() expected an error but got nil")
+				}
+
+				if g, w := err.Error(), "failed to create librarian state file"; !strings.Contains(g, w) {
+					t.Errorf("got %q, wanted it to contain %q", g, w)
+				}
+				return
+			}
+
+			if test.name == "invalid content parser" {
+				filePath := filepath.Join(tempDir, "state.yaml")
+				invalidContentParser := func(state *config.LibrarianState) ([]byte, error) {
+					return nil, errors.New("simulated parsing error")
+				}
+				err := writeLibrarianState(invalidContentParser, test.state, filePath)
+				if err == nil {
+					t.Errorf("writeLibrarianState() expected an error but got nil")
+				}
+
+				if g, w := err.Error(), "failed to convert state to bytes"; !strings.Contains(g, w) {
+					t.Errorf("got %q, wanted it to contain %q", g, w)
+				}
+				return
+			}
+
+			filePath := filepath.Join(tempDir, "state.yaml")
+			err := writeLibrarianState(contentParser, test.state, filePath)
+
+			if err != nil {
+				t.Fatalf("writeLibrarianState() unexpected error: %v", err)
+			}
+
+			// Verify the file content
+			gotBytes, err := os.ReadFile(filePath)
+			if err != nil {
+				t.Fatalf("Failed to read generated file: %v", err)
+			}
+
+			fileName := fmt.Sprintf("%s.yaml", test.name)
+			wantBytes, readErr := os.ReadFile(filepath.Join("..", "..", "testdata", fileName))
+			if readErr != nil {
+				t.Fatalf("Failed to read expected state for comparison: %v", readErr)
+			}
+
+			if diff := cmp.Diff(string(wantBytes), string(gotBytes)); diff != "" {
+				t.Errorf("Generated YAML mismatch (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func copyFile(dst, src string) (err error) {
+	sourceFile, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer sourceFile.Close()
+
+	destinationFile, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+
+	defer func() {
+		if err = errors.Join(err, destinationFile.Close()); err != nil {
+			err = fmt.Errorf("copyFile(%q, %q): %w", dst, src, err)
+		}
+	}()
+
+	_, err = io.Copy(destinationFile, sourceFile)
+
+	return err
 }
